requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.0
colorama==0.4.4
Flask==2.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
numpy==1.22.3
pandas==1.5.1
pycryptodome==3.15.0
pygubu==0.14
python-dateutil==2.8.2
pytz==2022.5
requests==2.28.1
six==1.16.0
urllib3==1.26.13
Werkzeug==2.1.0
